requirements.txt
Django==1.8.2
django-role-permissions
psycopg2
pyyaml
mixer
